+++ Changelog.md
@@ -1,6 +1,7 @@
 # FiReS
 
 ## git master
+* [!20] Check if properties exist for tasks Sort, Filer and Aggregate.
 * [!19] Added casters for enum properties.
 * [!18] Fixed Visual Studio 14 compilation.
 * [!17] Added property gid getters in PropertyManager